requirements.txt
Pillow==8.0.1	
numpy==1.19.2
pandas==1.1.4
scikit-learn==0.23.2	
streamlit==0.70.0

